#352 New partial for lead show
